examples/dtc/loadgen/pyproject.toml
[tool.poetry]
name = "moniterialize-loadgen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "2.0.*"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
ipykernel = "^6.13.0"
mypy = "^0.950"
types-psycopg2 = "^2.9.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
